+++ 수업필기_과제/3개월/API시연회/1212시작/선생님코드기반/DefaultWindow/CNormalShovel.cpp
@@ -38,7 +38,7 @@ int CNormalShovel::Update()
         Find_Player_OnSameTile();
 
         m_eRender = RENDER_MAPITEM;
-        m_iFrameTicker += m_fSpeed;
+        m_iFrameTicker += (int)m_fSpeed;
 
         switch (m_iFrameTicker)
         {
@@ -46,7 +46,7 @@ int CNormalShovel::Update()
         case 40:
         case 60:
         case 80:
-            m_iAnimationOffsetY += m_fSpeed;
+            m_iAnimationOffsetY += (int)m_fSpeed;
             break;
         case 0:
         case 100:
