+++ 수업필기_과제/3개월/API시연회/1212시작/선생님코드기반/DefaultWindow/CPlayerHP.cpp
@@ -45,45 +45,45 @@ void CPlayerHP::Render(HDC hDC)
         for (int i = 0; i < iLessHP / 2; ++i)
         {
             GdiTransparentBlt(hDC,					// 복사 받을 DC
-                m_tRect.left - 50.f * i,			// 복사 받을 위치 좌표 X, Y	
+                m_tRect.left - 50 * i,			// 복사 받을 위치 좌표 X, Y	
                 m_tRect.top,
-                m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 hMemDCEmpty,								// 복사할 이미지 DC	
                 0,	// 비트맵 출력 시작 좌표(Left, top)
                 0,
-                m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 RGB(255, 0, 255));					// 제거할 색상
             if (i + 1 == iLessHP / 2)
             {
-                m_tRect.left -= 50.f * (i + 1);
+                m_tRect.left -= 50 * (i + 1);
             }
         }
         GdiTransparentBlt(hDC,					// 복사 받을 DC
             m_tRect.left,			// 복사 받을 위치 좌표 X, Y	
             m_tRect.top,
-            m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
-            m_tInfo.fCY,
+            (int)m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
+            (int)m_tInfo.fCY,
             hMemDCHalf,								// 복사할 이미지 DC	
             0,	// 비트맵 출력 시작 좌표(Left, top)
             0,
-            m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
-            m_tInfo.fCY,
+            (int)m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
+            (int)m_tInfo.fCY,
             RGB(255, 0, 255));
-        m_tRect.left -= 50.f;
+        m_tRect.left -= 50;
         for (int i = 0; i < (m_iMaxHP - iLessHP) / 2; ++i)
         {
             GdiTransparentBlt(hDC,					// 복사 받을 DC
-                m_tRect.left - 50.f * i,			// 복사 받을 위치 좌표 X, Y	
+                m_tRect.left - 50 * i,			// 복사 받을 위치 좌표 X, Y	
                 m_tRect.top,
-                m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 hMemDCFull,								// 복사할 이미지 DC	
                 0,	// 비트맵 출력 시작 좌표(Left, top)
                 0,
-                m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 RGB(255, 0, 255));					// 제거할 색상
         }
     }
@@ -92,33 +92,33 @@ void CPlayerHP::Render(HDC hDC)
         for (int i = 0; i < iLessHP / 2; ++i)
         {
             GdiTransparentBlt(hDC,					// 복사 받을 DC
-                m_tRect.left - 50.f * i,			// 복사 받을 위치 좌표 X, Y	
+                m_tRect.left - 50 * i,			// 복사 받을 위치 좌표 X, Y	
                 m_tRect.top,
-                m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 hMemDCEmpty,								// 복사할 이미지 DC	
                 0,	// 비트맵 출력 시작 좌표(Left, top)
                 0,
-                m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 RGB(255, 0, 255));					// 제거할 색상
             if (i + 1 == iLessHP / 2)
             {
-                m_tRect.left -= 50.f * (i + 1);
+                m_tRect.left -= 50 * (i + 1);
             }
         }
         for (int i = 0; i < (m_iMaxHP - iLessHP) / 2; ++i)
         {
             GdiTransparentBlt(hDC,					// 복사 받을 DC
-                m_tRect.left - 50.f * i,			// 복사 받을 위치 좌표 X, Y	
+                m_tRect.left - 50 * i,			// 복사 받을 위치 좌표 X, Y	
                 m_tRect.top,
-                m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 hMemDCFull,								// 복사할 이미지 DC	
                 0,	// 비트맵 출력 시작 좌표(Left, top)
                 0,
-                m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
-                m_tInfo.fCY,
+                (int)m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
+                (int)m_tInfo.fCY,
                 RGB(255, 0, 255));					// 제거할 색상
         }
     }
