+++ 수업필기_과제/3개월/API시연회/1212시작/선생님코드기반/DefaultWindow/CPlayerWeaponUI.cpp
@@ -41,25 +41,25 @@ void CPlayerWeaponUI::Render(HDC hDC)
     GdiTransparentBlt(hDC,					// 복사 받을 DC
         m_tRect.left,			// 복사 받을 위치 좌표 X, Y	
         m_tRect.top,
-        m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
-        m_tInfo.fCY,
+        (int)m_tInfo.fCX,							// 복사 받을 이미지의 가로, 세로
+        (int)m_tInfo.fCY,
         hMemDC,								// 복사할 이미지 DC	
         0,	// 비트맵 출력 시작 좌표(Left, top)
         0,
-        m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
-        m_tInfo.fCY,
+        (int)m_tInfo.fCX,							// 복사할 이미지의 가로, 세로
+        (int)m_tInfo.fCY,
         RGB(255, 0, 255));					// 제거할 색상
 
     GdiTransparentBlt(hDC,					// 복사 받을 DC
-        m_tRect.left + (m_tInfo.fCX - m_pCurWeapon->Get_Info().fCX)/2,			// 복사 받을 위치 좌표 X, Y	
-        m_tRect.top + (m_tInfo.fCY - m_pCurWeapon->Get_Info().fCY) / 2 + 6,
-        m_pCurWeapon->Get_Info().fCX,							// 복사 받을 이미지의 가로, 세로
-        m_pCurWeapon->Get_Info().fCY,
+        m_tRect.left + ((int)m_tInfo.fCX - (int)m_pCurWeapon->Get_Info().fCX)/2,			// 복사 받을 위치 좌표 X, Y	
+        m_tRect.top + ((int)m_tInfo.fCY - (int)m_pCurWeapon->Get_Info().fCY) / 2 + 6,
+        (int)m_pCurWeapon->Get_Info().fCX,							// 복사 받을 이미지의 가로, 세로
+        (int)m_pCurWeapon->Get_Info().fCY,
         hMemDCCurShov,								// 복사할 이미지 DC	
         m_iItemPosX,	// 비트맵 출력 시작 좌표(Left, top)
         0,
-        m_pCurWeapon->Get_Info().fCX,							// 복사할 이미지의 가로, 세로
-        m_pCurWeapon->Get_Info().fCY,
+        (int)m_pCurWeapon->Get_Info().fCX,							// 복사할 이미지의 가로, 세로
+        (int)m_pCurWeapon->Get_Info().fCY,
         RGB(255, 0, 255));					// 제거할 색상
 
 }
