+++ 수업필기_과제/3개월/API시연회/1212시작/선생님코드기반/DefaultWindow/CTitaniumShovel.cpp
@@ -37,7 +37,7 @@ int CTitaniumShovel::Update()
         Find_Player_OnSameTile();
 
         m_eRender = RENDER_MAPITEM;
-        m_iFrameTicker += m_fSpeed;
+        m_iFrameTicker += (int)m_fSpeed;
 
         switch (m_iFrameTicker)
         {
@@ -45,7 +45,7 @@ int CTitaniumShovel::Update()
         case 40:
         case 60:
         case 80:
-            m_iAnimationOffsetY += m_fSpeed;
+            m_iAnimationOffsetY += (int)m_fSpeed;
             break;
         case 0:
